Consolidate title parsers
